feat: add validations for Quotation and POS Invoice; commonify
